fix(runtime): fix 12 bugs in runtime libraries and codegen
Fix bugs discovered during viperlib documentation audit:

- Fix rt_str_eq to return i1 (bool) instead of i64, with cascading
  updates across BASIC/Zia frontends, IL examples, and golden tests
- Fix Zia string != operator to zext i1 before icmp_eq comparison
- Fix CSV FormatLine to handle both raw and boxed strings
- Fix JsonPath and TOML parsers to detect boxed strings (RT_BOX_STR)
  in addition to raw RT_STRING_MAGIC strings
- Fix rtgen ilTypeToZiaType to handle "bool" return type mapping
- Fix rtgen cross-namespace return type annotations (Crypto->Bytes,
  Csv/Markdown/Html->Seq) via explicit override map
- Add Seq.get_Count alias in runtime.def
- Fix ARM64 codegen for Zext1 and Trunc1 opcodes
- Fix x86_64 Trunc1 to use movzx for proper zero-extension
- Fix -Wstring-compare warning in LowerExprNumeric.cpp
- Add regression tests for string, text, and crypto bug fixes
- Update viperlib documentation with Zia examples and corrections

tests: 3 new Zia runtime regression test suites (string, text, crypto),
ARM64 codegen bugfix tests, updated golden tests and unit tests

+++ src/runtime/rt_string.h
@@ -216,7 +216,7 @@ extern "C"
     /// @param a First operand.
     /// @param b Second operand.
     /// @return 1 when equal, 0 otherwise.
-    int64_t rt_str_eq(rt_string a, rt_string b);
+    int8_t rt_str_eq(rt_string a, rt_string b);
 
     /// @brief Compare @p a < @p b lexicographically.
     /// @param a First operand.
